Replace main service by container itself
