+++ service/graphql/config.go
@@ -1,7 +1,7 @@
 package graphql
 
 import (
-	"github.com/Sanchous98/project-confucius-backend"
+	"github.com/Sanchous98/project-confucius-backend/utils"
 	"os"
 )
 
@@ -10,7 +10,7 @@ type Config struct {
 }
 
 func (gqlc *Config) HydrateConfig() error {
-	config, err := confucius.HydrateConfig(gqlc, os.Getenv("CONFIG_PATH")+"/graphql.yml")
+	config, err := utils.HydrateConfig(gqlc, os.Getenv("CONFIG_PATH")+"/graphql.yml")
 
 	if err != nil {
 		return err
